+++ extensions/rebaptize/src/preset-sequential.tsx
@@ -1,4 +1,0 @@
-import Rebaptize from "./rebaptize";
-export default function () {
-  return <Rebaptize initialMode="sequential" />;
-}